Consolidate 2 BaseSparkTest classes; make SparkContext handling more robust (avoid RCs and multiple SCs running at once)
