+++ datavec-spark/src/test/java/org/datavec/spark/transform/analysis/TestAnalysis.java
@@ -21,8 +21,8 @@ import org.datavec.api.transform.analysis.DataAnalysis;
 import org.datavec.api.transform.analysis.columns.*;
 import org.datavec.api.transform.schema.Schema;
 import org.datavec.api.writable.*;
+import org.datavec.spark.BaseSparkTest;
 import org.datavec.spark.transform.AnalyzeSpark;
-import org.datavec.spark.transform.BaseSparkTest;
 import org.apache.spark.api.java.JavaRDD;
 import org.joda.time.DateTimeZone;
 import org.junit.Test;
